Changed a lot and forgot to commit. More info in description.
Totally forgot to commit, so this is all the changes done in the last month.

Serverside:

Added functions for finding players, and the game they're in.

Created a question generator script

Made player object have the question, answer, and when they submitted.

Client Side:
Added the question page and it is based off of the server.
Added A waiting Screen once a player has answered the question (not connected to server)
Added incorrect or correct screen (not connected to server)
Added leaderboard screen (not connected to server).

I'm closing in on the end of this projecy, just need to type it all to the server, fix the client side error handling and than do some testing.