Finished game loop and added quality of life buttons.
Changed options for max players from 3 and 4 to 10 and 25. (I'm not sure how the server would handle 25 ppl but why not? :)

This project is almost complete, all I need to do is bug fixes and quality of life upgrades.

Known Bugs:
If everyone answers wrong then the last to answer gets it correct. (Server side Issue)

The join page ever so often loads multiple of the same game. (client side issue).

Answer buttons for a question ever so often load the same number for multiple buttons (server side).